requirements.txt
# Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv

# Base Vectorial
chromadb
PyMuPDF
Pillow
python-jose[cryptography]
bcrypt
python-multipart
# IA / LLM
google-generativeai

# Procesamiento de documentos
pymupdf
pillow
langchain
langchain-text-splitters

# Frontend
streamlit
requests
